Back End/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
kaggle==1.5.16
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
joblib==1.3.2
